server/requirements.txt
Django==4.2.21
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.8.0
django-filter==25.1
gunicorn==22.0.0
whitenoise==6.7.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
